+++ src/inc/taosmsg.h
@@ -370,7 +370,7 @@ typedef struct SExprInfo {
   struct tExprNode* pExpr;
   int16_t     bytes;
   int16_t     type;
-  int16_t     interBytes;
+  int32_t     interBytes;
 } SExprInfo;
 
 typedef struct SColumnFilterInfo {
